+++ cmd/data-node/node/node.go
@@ -262,7 +262,7 @@ func (l *NodeCommand) runNode(args []string) error {
 
 	// start gateway
 	if l.conf.GatewayEnabled {
-		gty := server.New(l.conf.Gateway, l.Log)
+		gty := server.New(l.conf.Gateway, l.Log, l.vegaPaths)
 
 		eg.Go(func() error { return gty.Start(ctx) })
 	}